+++ src/types/plugins/replication.d.ts
@@ -18,7 +18,7 @@ export type InternalStoreReplicationPullDocType<RxDocType> = InternalStoreDocTyp
 }>;
 
 export type ReplicationPullHandlerResult<RxDocType, CheckpointType> = {
-    checkpoint: CheckpointType | null;
+    checkpoint: CheckpointType | undefined;
     documents: WithDeleted<RxDocType>[];
 };
 
@@ -28,6 +28,7 @@ export type ReplicationPullHandler<RxDocType, CheckpointType> = (
     lastPulledCheckpoint: CheckpointType | undefined,
     batchSize: number
 ) => Promise<ReplicationPullHandlerResult<RxDocType, CheckpointType>>;
+
 export type ReplicationPullOptions<RxDocType, CheckpointType> = {
     /**
      * A handler that pulls the new remote changes