+++ test/unit/key-compression.test.ts
@@ -29,9 +29,10 @@ import {
     enableKeyCompression,
     ensureReplicationHasNoErrors,
     HumanDocumentType,
-    human
+    human,
+    getPullHandler,
+    getPushHandler
 } from '../../plugins/test-utils/index.mjs';
-import { getPullHandler, getPushHandler } from './replication.test.ts';
 import { replicateRxCollection } from '../../plugins/replication/index.mjs';
 import type { SimpleHumanDocumentType } from '../../src/plugins/test-utils/schema-objects.ts';